% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACViewTrack.R
\name{addGeneNameToTxdb}
\alias{addGeneNameToTxdb}
\title{Add Gene sybmol to the GRanges object after calling genes(txdb)}
\usage{
addGeneNameToTxdb(
  txdb = TxDb.Hsapiens.UCSC.hg19.knownGene,
  eg.db = org.Hs.eg.db
)
}
\arguments{
\item{txdb}{A TxDb object, e.g. TxDb.Hsapiens.UCSC.hg19.knownGene}

\item{eg.db}{Either org.Hs.eg.db or org.Mm.eg.db}
}
\value{
An GRanges object for all the genes with gene symbol in the metadata column
}
\description{
Add Gene sybmol to the GRanges object after calling genes(txdb)
}
