% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACAnnotateCelltype.R
\docType{methods}
\name{ATACAnnotateCelltype}
\alias{ATACAnnotateCelltype}
\title{Annotate scATAC-seq celltype based on marker genes}
\usage{
ATACAnnotateCelltype(
  ATAC,
  signatures = "human.immune.CIBERSORT",
  min.score = 0,
  genes.test.use = "presto",
  genes.cutoff = 1e-05,
  orig.ident = NULL
)
}
\arguments{
\item{ATAC}{Seurat object of clustered scATAC-seq dataset, generated by \code{\link{ATACRunSeurat}} function.}

\item{signatures}{The \code{signatures} can be character or data frame. If \code{signatures} is character, 
it should be "human.immune.CIBERSORT", "mouce.brain.ALLEN", "mouse.all.facs.TabulaMuris" or "mouse.all.droplet.TabulaMuris".
If \code{signatures} is a data frame, is should be a data frame of celltype signatures, with first column is celltype, and second column is
gene symbol. Check the immune signatures from CIBERSORT (Newman et al., Nature Method, 2015) in example for details. Default is "human.immune.CIBERSORT".}

\item{min.score}{Minimum score required. For one cluster, if the score for all celltypes are less than 
\code{min.score}, the cluster will be annotated as "Others". Default is 0.}

\item{genes.test.use}{Denotes which test to use to identify genes. Default is "presto".}

\item{genes.cutoff}{Identify differential expressed genes with adjusted p.value less than \code{genes.cutoff} as cluster speficic genes
for each cluster. Default cutoff is 1E-5.}

\item{orig.ident}{Orignal identity. If given, will use the original identity to annotate the celltypes. Default is NULL.}
}
\value{
An ATAC Seurat object with celltype annotated, and the default assay have been set to gene activity.
}
\description{
Annotate the celltype for all clusters based on the marker genes and the input celltype signatures for scATAC-seq dataset.
}
\examples{
data(pbmc.ATAC)
data(pbmc.RP)
data(human.immune.CIBERSORT)
pbmc.ATAC.res <- ATACRunSeurat(inputMat = pbmc.ATAC, project = "PBMC.scATAC.Seurat", method = "LSI")
pbmc.ATAC.res$ATAC <- ATACAttachGenescore(pbmc.ATAC.res$ATAC, pbmc.RP)
pbmc.ATAC.res$ATAC <- ATACAnnotateCelltype(pbmc.ATAC.res$ATAC, human.immune.CIBERSORT, min.score = 0.1, genes.cutoff = 1E-3)
str(pbmc.ATAC.res$ATAC)

}
\author{
Chenfei Wang, Dongqing Sun
}
