% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACAnnotateTranscriptionFactor.R
\docType{methods}
\name{ATACAnnotateTranscriptionFactor}
\alias{ATACAnnotateTranscriptionFactor}
\title{Identify driver transcrition factors for scATAC-seq clusters}
\usage{
ATACAnnotateTranscriptionFactor(ATAC, peaks, project = ATAC@project.name,
  giggle.path, organism = "GRCh38", top.tf = 10)
}
\arguments{
\item{ATAC}{Seurat object of ATAC RPscore clustering  generated by \code{\link{ATACAnnotateCelltype}} function.}

\item{peaks}{Data frame of differential peaks, generated by \code{\link{ATACRunSeurat}} function.}

\item{project}{Output project name for the giggle analysis result. Default is the project name of the ATAC Seurat object}

\item{organism}{Organism for the dataset. Only support "GRCh38" and "GRCm38". Default is "GRCh38".}

\item{top.tf}{Output the target genes for the top TFs. Default is 10, only output the target genes for the top 10 TFs.}

\item{giggle.anno.path}{Path of the giggle annotation files. The giggle annoation for CistromeDB is available at https://github.com/liulab-dfci/MAESTRO.
Download and unzip the annotation, supply the path of the annotations to \code{giggle.anno.path}.}
}
\value{
A list of top enriched transcription factors and their target genes for each cluster.
}
\description{
Identify driver transcrition factors for scATAC-seq clusters based on TF ChIP-seq peaks from CistromeDB. To run this function, you must first install the giggle package (see from https://github.com/ryanlayer/giggle).
}
\examples{
data(pbmc.ATAC)
data(pbmc.RP)
data(human.immune.CIBERSORT)
pbmc.ATAC.res <- ATACRunSeurat(inputMat = pbmc.ATAC, project = "PBMC.scATAC.Seurat", method = "LSI")
pbmc.ATAC.res$ATAC <- ATACAnnotateCelltype(pbmc.ATAC.res$ATAC, pbmc.RP, human.immune.CIBERSORT, min.score = 0.1, genes.cutoff = 1E-3)
pbmc.tfs <- ATACAnnotateTranscriptionFactor(ATAC = pbmc.ATAC.res$ATAC, peaks = pbmc.ATAC.res$peaks, project = "PBMC.scATAC.TF", giggle.path = "/home/annotations/giggle")
pbmc.tfs

}
\author{
Dongqing Sun, Changxin Wan, Chenfei Wang
}
