% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACAttachGenescore.R
\docType{methods}
\name{ATACAttachGenescore}
\alias{ATACAttachGenescore}
\title{Add gene score (gene activity) matrix to an ATAC Seurat object}
\usage{
ATACAttachGenescore(ATAC, RPmatrix)
}
\arguments{
\item{ATAC}{Seurat object of clustered scATAC-seq dataset, generated by \code{\link{ATACRunSeurat}} function.}

\item{RPmatrix}{Regulatory potential matrix generated by \code{\link{ATACCalculateGenescore}}. 
With genes as rows and cells as columns, and gene RP score as values.}
}
\value{
An ATAC Seurat object with ACTIVITY attached.
}
\description{
Add gene score (gene activity) matrix to an ATAC Seurat object and normalize it.
}
\examples{
data(pbmc.ATAC)
data(pbmc.RP)
pbmc.ATAC.res <- ATACRunSeurat(inputMat = pbmc.ATAC, project = "PBMC.scATAC.Seurat", method = "LSI")
pbmc.ATAC.res$ATAC <- ATACAttachGenescore(pbmc.ATAC.res$ATAC, pbmc.RP)
str(pbmc.ATAC.res$ATAC)

}
\author{
Chenfei Wang, Dongqing Sun
}
