% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACCalculateGenescore.R
\docType{methods}
\name{ATACCalculateGenescore}
\alias{ATACCalculateGenescore}
\title{Estimate gene activity from peak using regulatory potential model}
\usage{
ATACCalculateGenescore(
  inputMat,
  project = "MAESTRO.scATAC",
  organism = "GRCh38"
)
}
\arguments{
\item{inputMat}{Input unnormalized count matrix, with peaks as rows and cells as columns. Rownames should be like "chromosome_peakstart_peakend",
for example "chr10_100020591_100020841". It can be a sparse matrix or a dense matrix. But we recommend to use the sparse one, because 
this will result in significant memory and speed savings.}

\item{project}{Output project name for the gene sore result.Default is "MAESTRO.scATAC".}

\item{organism}{Organism for the dataset. Only support "GRCh38" and "GRCm38". Default is "GRCh38".}
}
\value{
Sparse matrix of regulatory potential generated by MAESTRO. With genes as rows and cells as columns,
and gene RP score as values.
}
\description{
Calculate gene score according to scATAC peak count.
}
\examples{
data(pbmc.ATAC)
pbmc.ATAC.RP.res <- ATACCalculateGenescore(inputMat = pbmc.ATAC)

}
\author{
Dongqing Sun, Changxin Wan, Chenfei Wang
}
