% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACFragmentSizePlot.R
\docType{methods}
\name{ATACFragmentSizePlot}
\alias{ATACFragmentSizePlot}
\title{Fragment size distribution plot for scATAC-seq analysis}
\usage{
ATACFragmentSizePlot(filepath, name)
}
\arguments{
\item{filepath}{Path of the fragment size file generated by \code{MAESTRO}.}

\item{name}{Name for the output fragment size distribution plot.}
}
\description{
Generate the fragment size distribution plot for scATAC-seq quality control analysis.
}
\author{
Chenfei Wang, Dongqing Sun
}
