% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACMapPlot.R
\docType{methods}
\name{ATACMapPlot}
\alias{ATACMapPlot}
\title{Mapping summary plot for scATAC-seq analysis}
\usage{
ATACMapPlot(filepath, platform = "10x-genomics", name)
}
\arguments{
\item{filepath}{Path of the scATAC-seq mapping summary file generated by \code{MAESTRO}.}

\item{platform}{Denote platform of the scATAC-seq data, available options are "microfluidic",
"10x-genomics","sci-ATAC-seq". Default platform is "10x-genomics".}

\item{name}{Name for the output mapping summary plot.}
}
\description{
Summary plot of total reads, mappability, duplicate reads ratio, mitochondria reads ratio for scATAC-seq analysis.
}
\author{
Chenfei Wang, Dongqing Sun
}
