% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACReadDistrPlot.R
\docType{methods}
\name{ATACReadDistrPlot}
\alias{ATACReadDistrPlot}
\title{Read distribution plot for scATAC-seq analysis}
\usage{
ATACReadDistrPlot(stat.filepath, name, platform)
}
\arguments{
\item{stat.filepath}{Path of the single-cell statistic file generated by \code{MAESTRO}.}

\item{name}{Name for the output read distribution plot.}

\item{platform}{Denote platform of the scATAC-seq data, available options are "microfluidic",
"10x-genomics","sci-ATAC-seq". Default platform is "10x-genomics".}
}
\description{
Summary plot of total reads, duplicate reads ratio, mappability, mitochondria and peak reads ratio for scATAC-seq analysis.
}
\author{
Chenfei Wang, Dongqing Sun
}
