% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACRunchromVAR.R
\docType{methods}
\name{ATACRunchromVAR}
\alias{ATACRunchromVAR}
\title{Evaluate the transcription factor enrichment in scATAC-seq using chromVAR}
\usage{
ATACRunchromVAR(
  inputMat,
  project = "MAESTRO.scATAC.chromVAR",
  min.c = 50,
  min.p = 500,
  organism = "GRCh38"
)
}
\arguments{
\item{inputMat}{Input unnormalized count matrix, with peaks as rows and cells as columns. Rownames should be like "chromosome_peakstart_peakend",
for example "chr10_100020591_100020841".}

\item{project}{Output project name for the chromVAR object. Default is "MAESTRO.scATAC.chromVAR".}

\item{min.c}{Minimum number of cells required for a peak. Will exclude the peaks from input matrix if they only identified in 
less than \code{min.c} cells. Default is 50.}

\item{min.p}{Minimum number of peaks required for a cell. Will exclude the cells from input matrix if less than \code{min.p}
peaks are deteced in one cell. Default is 500.}

\item{organism}{Organism for the dataset. Only support "GRCh38" and "GRCm38". Default is "GRCh38".}
}
\value{
A list contains a chromVAR deviation object, and a data frame with rows as TFs, columns as cells, and values as TF z-score in each cell.
}
\description{
Evaluate the transcription factor enrichment in scATAC-seq using chromVAR (version >= 1.1.1).
}
\examples{
data(pbmc.ATAC)
pbmc.ATAC.res <- ATACRunchromVAR(inputMat = pbmc.ATAC, project = "PBMC.scATAC.chromVAR")
str(pbmc.ATAC.res$dev)
head(pbmc.ATAC.res$zscore[,1:10])

}
\author{
Chenfei Wang
}
