% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACRuncicero.R
\docType{methods}
\name{ATACRuncicero}
\alias{ATACRuncicero}
\title{Generate gene score matrix for scATAC-seq dataset using cicero}
\usage{
ATACRuncicero(peakCountFile, outPrefix)
}
\arguments{
\item{inputMat}{Input unnormalized count matrix, with three columns, peak, cell and count. Peaks should be like "chromosome_peakstart_peakend",
for example "chr10_100020591_100020841".}

\item{project}{Output project name for the cicero object. Default is "MAESTRO.scATAC.cicero".}
}
\value{
A list contains a cicero object for ATAC clusters and a list with clustering informations.
}
\description{
Generate gene score matrix for scATAC-seq dataset using cicero.
}
\examples{
data(pbmc.ATAC)
pbmc.ATAC.res <- ATACRuncicero(inputMat = pbmc.ATAC, project = "PBMC.scATAC.cicero")
str(pbmc.ATAC.res$ATAC)
head(pbmc.ATAC.res$cluster)

}
\author{
Chenfei Wang
}
