% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACRunscABC.R
\docType{methods}
\name{ATACRunscABC}
\alias{ATACRunscABC}
\title{Clustering analysis for scATAC-seq data using scABC}
\usage{
ATACRunscABC(
  inputMat,
  project = "MAESTRO.scATAC.scABC",
  min.c = 50,
  min.p = 500,
  cluster.number = 10
)
}
\arguments{
\item{inputMat}{Input unnormalized count matrix, with peaks as rows and cells as columns. Rownames should be like "chromosome_peakstart_peakend",
for example "chr10_100020591_100020841".}

\item{project}{Output project name. Default is "MAESTRO.scATAC.scABC".}

\item{min.c}{Minimum number of cells required for a peak. Will exclude the peaks from input matrix if they only identified in 
less than \code{min.c} cells. Default is 50.}

\item{min.p}{Minimum number of peaks required for a cell. Will exclude the cells from input matrix if less than \code{min.p}
peaks are deteced in one cell. Default is 500.}

\item{cluster.number}{Number of K for K-medoid clustering analysis. Default is 10.}
}
\value{
A list with clustering informations.
}
\description{
Clustering analysis for scATAC-seq dataset using scABC. Including normalization, scABC clustering and UMAP visualization. To run UMAP analysis, you must first install the umap-learn python package (e.g. via \code{pip install umap-learn}).
}
\examples{
data(pbmc.ATAC)
pbmc.ATAC.res <- ATACRunscABC(inputMat = pbmc.ATAC, project = "PBMC.scATAC.scABC")
head(pbmc.ATAC.res)

}
\author{
Chenfei Wang
}
