% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cli.R
\name{baton_parse_cli}
\alias{baton_parse_cli}
\title{cli output for generate schedule table}
\usage{
baton_parse_cli(parse_succeeds, parse_errors)
}
\arguments{
\item{parse_succeeds}{list of parse results (i.e., succeeded)}

\item{parse_errors}{list of parse errors}
}
\value{
cli output
}
\description{
cli output for generate schedule table
}
