% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_schedule_entry.R
\name{build_schedule_entry}
\alias{build_schedule_entry}
\title{Parse and validate tags then create and populate MaestroPipelineList}
\usage{
build_schedule_entry(script_path)
}
\arguments{
\item{script_path}{path to script}
}
\value{
MaestroPipelineList R6 class
}
\description{
Parse and validate tags then create and populate MaestroPipelineList
}
\keyword{internal}
