% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_seconds}
\alias{convert_to_seconds}
\title{Convert a duration string to number of seconds}
\usage{
convert_to_seconds(time_string)
}
\arguments{
\item{time_string}{string like 1 day, 2 weeks, 12 hours, etc.}
}
\value{
number of seconds
}
\description{
Convert a duration string to number of seconds
}
\keyword{internal}
