% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_orchestrator.R
\name{create_orchestrator}
\alias{create_orchestrator}
\title{Create a new orchestrator}
\usage{
create_orchestrator(
  path,
  type = c("R", "Quarto", "RMarkdown"),
  open = interactive(),
  quiet = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{path}{file path for the orchestrator script}

\item{type}{file type for the orchestrator (supports R, Quarto, and RMarkdown)}

\item{open}{whether or not to open the script upon creation}

\item{quiet}{whether to silence messages in the console (default = \code{FALSE})}

\item{overwrite}{whether to overwrite an existing orchestrator or maestro project}
}
\value{
invisible
}
\description{
Create a new orchestrator
}
