% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_check.R
\name{dag_validity_check}
\alias{dag_validity_check}
\title{Checks the validity of DAG components of a schedule}
\usage{
dag_validity_check(schedule)
}
\arguments{
\item{schedule}{a schedule table returned from \code{build_schedule}}
}
\value{
invisible or error
}
\description{
Checks the validity of DAG components of a schedule
}
