% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.run_sequence_days_out}
\alias{.run_sequence_days_out}
\title{Number of days ahead to generate run sequences for a given frequency unit
Used by the observable get_run_sequence path.}
\usage{
.run_sequence_days_out(unit)
}
\arguments{
\item{unit}{character frequency unit}
}
\value{
integer
}
\description{
Number of days ahead to generate run sequences for a given frequency unit
Used by the observable get_run_sequence path.
}
\keyword{internal}
