% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACViewTrack.R
\name{extend}
\alias{extend}
\title{Extend a GRanges object upstream and downstream}
\usage{
extend(x, upstream = 0, downstream = 0)
}
\arguments{
\item{x}{A GRanges object}

\item{upstream}{bp for extending upstream}

\item{downstream}{bp for extending downstream}
}
\value{
An extended GRanges object
}
\description{
Extend a GRanges object upstream and downstream
}
