% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindAllMarkersMAESTRO.R
\docType{methods}
\name{FindAllMarkersMAESTRO}
\alias{FindAllMarkersMAESTRO}
\title{Identify differential features of single-cell transcriptome and regulome data for each of the clusters}
\usage{
FindAllMarkersMAESTRO(
  object,
  test.use = "presto",
  features = NULL,
  min.pct = 0.1,
  logfc.threshold = 0.25,
  only.pos = FALSE,
  verbose = TRUE,
  return.thresh = 0.01,
  min.cells.feature = 3,
  min.cells.group = 3,
  slot = "data",
  latent.vars = NULL
)
}
\arguments{
\item{object}{Seurat object generated by \code{\link{RNARunSeurat}} function.}

\item{test.use}{Method to use to identify differential genes or peaks. Default is "presto", a fast version of Wilcoxon Rank Sum test.
"presto" produces exactly the same result as "wilcox", but "presto" is much faster.
For scATAC-seq, "t" or "wilcox", is another option. For scRNA-seq, other available options are "wilcox", "bimod", 
"roc", "t", "negbinom", "poisson", "LR", "MAST", "DESeq2", which supported by Seurat \code{link{FindAllMarkers}} function.}

\item{features}{Features to use. Default is to use all features.}

\item{min.pct}{Only test features that are detected in a minimum fraction of min.pct cells 
in either of the two populations. For genes, default is 0.1.}

\item{logfc.threshold}{Only test features which show X-fold difference (log-scale)
between two group of cells. For genes, default is 0.25.}

\item{only.pos}{Only return positive markers, default is FALSE.}

\item{verbose}{Print a progress bar, default is TRUE.}

\item{return.thresh}{Only return markers that have a p-value < return.thresh, 
or a power > return.thresh (if the test is ROC)}

\item{min.cells.feature}{Minimum number of cells expressing the feature in at least one of the two groups, 
currently only used for poisson and negative binomial tests. Default is 3.}

\item{min.cells.group}{Minimum number of cells in one of the groups. Default is 3.}

\item{slot}{Slot to pull data from; note that if test.use is "negbinom", "poisson", or "DESeq2", 
slot will be set to "counts". Default is "data"}

\item{latent.vars}{Variables to test, used only when test.use is one of 'LR', 'negbinom', 'poisson', or 'MAST'}
}
\value{
A dataframe containing a ranked list of pytative markers and 
associated statics(p-value, ROC score, etc.)
}
\description{
Identify differential features of single-cell data (genes for transcriptome and peaks for regulome) for each of the clusters
}
\author{
Dongqing Sun, Chenfei Wang
}
