% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarkersMAESTRO.R
\docType{methods}
\name{FindMarkersMAESTRO}
\alias{FindMarkersMAESTRO}
\title{Identify differential features of single-cell transcriptome and regulome data for given cluster}
\usage{
FindMarkersMAESTRO(
  object,
  ident.1 = 0,
  ident.2 = NULL,
  test.use = "presto",
  slot = "data",
  features = NULL,
  min.pct = 0.1,
  logfc.threshold = 0.25,
  latent.vars = NULL,
  min.cells.feature = 3,
  min.cells.group = 3,
  only.pos = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object generated by \code{\link{CreateSeuratObject}} function.}

\item{ident.1}{The identy of interested cluster, default is 0.}

\item{ident.2}{A second identity of cluster for comparison; if \code{NULL},
use all other cells for comparison. Default is \code{NULL}.}

\item{test.use}{Method to use to identify differential genes or peaks. Default is "presto", a fast version of Wilcoxon Rank Sum test.
"presto" produces exactly the same result as "wilcox", but "presto" is much faster.
For scATAC-seq, "t" test is another option. For scRNA-seq, other available options are "bimod",
"roc", "t", "negbinom", "poisson", "LR", "MAST", "DESeq2", which supported by Seurat \code{link{FindMarkers}} function.}

\item{slot}{Slot to pull data from; note that if test.use is "negbinom", "poisson", or "DESeq2", 
slot will be set to "counts". Default is "data"}

\item{features}{Features to use. Default is to use all features.}

\item{min.pct}{Only test features that are detected in a minimum fraction of min.pct cells 
in either of the two populations. For genes, default is 0.1.}

\item{logfc.threshold}{Only test features which show X-fold difference (log-scale)
between two group of cells. For genes, default is 0.25.}

\item{latent.vars}{Variables to test, used only when test.use is one of 'LR', 'negbinom', 'poisson', or 'MAST'}

\item{min.cells.feature}{Minimum number of cells expressing the feature in at least one of the two groups, 
currently only used for poisson and negative binomial tests. Default is 3.}

\item{min.cells.group}{Minimum number of cells in one of the groups. Default is 3.}

\item{only.pos}{Only return positive markers, default is FALSE.}

\item{verbose}{Print a progress bar, default is TRUE.}
}
\value{
A dataframe containing a ranked list of pytative markers and 
associated statics(p-value, ROC score, etc.)
}
\description{
Identify differential features of single-cell data (genes for transcriptome and peaks for regulome) for given cluster. Modified from Seurat FindMarkers function.
}
\author{
Dongqing Sun, Chenfei Wang
}
