% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels.R
\name{get_labels}
\alias{get_labels}
\title{Get the labels of pipelines in a MaestroSchedule object}
\usage{
get_labels(schedule)
}
\arguments{
\item{schedule}{object of type MaestroSchedule created using \code{build_schedule()}}
}
\value{
data.frame
}
\description{
Creates a data.frame of labels for all labelled pipelines in the schedule.
}
\examples{
if (interactive()) {
  pipeline_dir <- tempdir()
  create_pipeline("my_new_pipeline", pipeline_dir, open = FALSE)
  schedule <- build_schedule(pipeline_dir = pipeline_dir)

  get_labels(schedule)

  # Alternatively, use the underlying R6 method
  schedule$get_labels()
}
}
