% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineage.R
\name{get_lineage}
\alias{get_lineage}
\title{Get the lineage of pipeline executions in a schedule run}
\usage{
get_lineage(schedule)
}
\arguments{
\item{schedule}{object of type MaestroSchedule created using \code{build_schedule()}}
}
\value{
data.frame
}
\description{
For DAG pipelines, lineage is the network connecting each pipeline run to its downstream pipeline run.
This function returns a from-to data.frame where each run_id is linked to its downstream run_id. This
function is particularly useful when paired with \code{get_status()} for the purpose of tracking statuses
of distinct pipeline lineages.
}
\examples{
if (interactive()) {
  pipeline_dir <- tempdir()
  create_pipeline("my_new_pipeline", pipeline_dir, open = FALSE)
  schedule <- build_schedule(pipeline_dir = pipeline_dir)

  get_lineage(schedule)

  # Alternatively, use the underlying R6 method
  schedule$get_lineage()
}
}
