% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{get_network}
\alias{get_network}
\title{Get the network structure of pipelines in a MaestroSchedule object}
\usage{
get_network(schedule)
}
\arguments{
\item{schedule}{object of type MaestroSchedule created using \code{build_schedule()}}
}
\value{
data.frame with columns \code{from} and \code{to}
}
\description{
Returns the pipeline dependency structure as an edge list data.frame.
Each row represents a directed dependency between two pipelines. The
result will be empty if there are no DAG pipelines in the schedule.
}
\examples{
if (interactive()) {
  pipeline_dir <- tempdir()
  create_pipeline("my_new_pipeline", pipeline_dir, open = FALSE)
  schedule <- build_schedule(pipeline_dir = pipeline_dir)

  get_network(schedule)

  # Alternatively, use the underlying R6 method
  schedule$get_network()
}
}
\seealso{
\code{\link[=show_network]{show_network()}} which is deprecated in favour of this function.
}
