% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pipeline_run_sequence}
\alias{get_pipeline_run_sequence}
\title{Generate a sequence of run times for a pipeline}
\usage{
get_pipeline_run_sequence(
  pipeline_n,
  pipeline_unit,
  pipeline_datetime,
  check_datetime
)
}
\arguments{
\item{pipeline_n}{number of units for the pipeline frequency}

\item{pipeline_unit}{unit for the pipeline frequency}

\item{pipeline_datetime}{datetime of the first time the pipeline is to run}

\item{check_datetime}{datetime against which to check the running of pipelines (default is current system time in UTC)}
}
\value{
vector of timestamps or dates
}
\description{
Generate a sequence of run times for a pipeline
}
