% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_pipelines.R
\name{identify_pipelines}
\alias{identify_pipelines}
\title{Identify pipelines}
\usage{
identify_pipelines(
  orch_interval,
  orch_unit,
  check_datetime = Sys.time(),
  pipeline_interval,
  pipeline_freq,
  pipeline_datetime
)
}
\arguments{
\item{orch_interval}{a numeric value representing how often the orchestration runs}

\item{orch_unit}{unit of how often the orchestration runs (i.e. min, hour, day, month, etc.)}

\item{check_datetime}{datetime to be used to check if the pipeline should run, default is system time}

\item{pipeline_interval}{a numeric value representing how often the pipeline runs}

\item{pipeline_freq}{unit of how often the pipeline runs (i.e. min, hour, day, month, etc.)}

\item{pipeline_datetime}{datatime of the first time the pipeline is to run}
}
\description{
Identify pipelines
}
