% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incorporate.R
\docType{methods}
\name{Incorporate}
\alias{Incorporate}
\title{Incorparation of scATAC-seq clusters with scRNA-seq clusters}
\usage{
Incorporate(
  RNA,
  ATAC,
  RPmatrix = NULL,
  project = "MAESTRO.coembedding",
  method = "MAESTRO",
  dims.use = 1:30,
  RNA.res = 0.6,
  ATAC.res = 0.6
)
}
\arguments{
\item{RNA}{Seurat object of clustered scRNA-seq dataset, generated by \code{\link{RNARunSeurat}} function.}

\item{ATAC}{Seurat object of clustered scATAC-seq dataset, generated by \code{\link{ATACRunSeurat}} function.}

\item{RPmatrix}{Data frame of regulatory potential matrix generated by MAESTRO. With genes as rows and cells as columns,
and gene RP score as values. Can be ignored if \code{\link{ATACAnnotateCelltype}} have already been run.}

\item{project}{Output project name. Default is "MAESTRO.coembedded".}

\item{method}{Method to do integration, MAESTRO or Seurat. If "MAESTRO", gene RP score will be used to quantify the gene activity for scATAC-seq.
If "Seurat" is set, \code{\link{CreateGeneActivityMatrix}} from Seurat will be used to model the gene activity.}

\item{dims.use}{Number of dimensions used for PCA and UMAP analysis. Default is 1:30, use the first 30 PCs.}

\item{RNA.res}{Clusterig resolution used for the scRNA-seq dataset, should keep the same with the input RNA object. Default is 0.6.}

\item{ATAC.res}{Clustering resolution used for the scATAC-seq dataset, should keep the sampe with the input ATAC object. Default is 0.6.}
}
\value{
A combined Seurat object with RNA dataset, ATAC dataset, gene activity dataset, combined UMAP analysis and clustering information. A tsv file for the cell meta information.
}
\description{
Incorparation of scATAC-seq clusters with scRNA-seq clusters, generate the cell-coembedding visualizations, also transfer the celltypes from scRNA-seq to scATAC-seq annotations.
}
\examples{
data(pbmc.RNA)
data(pbmc.ATAC)
data(pbmc.RP)

pbmc.RNA.res <- RNARunSeurat(inputMat = pbmc.RNA, project = "PBMC.scRNA.Seurat")
pbmc.RNA.res$RNA <- RNAAnnotateCelltype(pbmc.RNA.res$RNA, pbmc.RNA.res$genes, human.immune.CIBERSORT, min.score = 0.05)
pbmc.ATAC.res <- ATACRunSeurat(inputMat = pbmc.ATAC, project = "PBMC.scATAC.Seurat")
pbmc.ATAC.res$ATAC <- ATACAnnotateCelltype(pbmc.ATAC.res$ATAC, pbmc.RP, human.immune.CIBERSORT, min.score = 0.1, genes.cutoff = 1E-3)
pbmc.coembedded.cluster <- Incorporate(RNA = pbmc.RNA.res$RNA, ATAC = pbmc.ATAC.res$ATAC, project = "PBMC.coembedded")
str(pbmc.coembedded.cluster)

}
\author{
Chenfei Wang
}
