% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_dag}
\alias{is_valid_dag}
\title{Checks whether a DAG is valid (no loops)}
\usage{
is_valid_dag(edges)
}
\arguments{
\item{edges}{a data.frame of edges (from, to)}
}
\value{
boolean
}
\description{
Checks whether a DAG is valid (no loops)
}
\keyword{internal}
