% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNNSmoothing.R
\name{knn_smoothing}
\alias{knn_smoothing}
\title{kNN-smoothing on a given matrix based on a pre-calculated sNN matrix}
\usage{
knn_smoothing(M, SNN, k = 10, method = "ave", strict = TRUE)
}
\arguments{
\item{M}{A numeric matrix that will be smoothed. It should have
feature names on rows and cell/barcode names on columns.}

\item{SNN}{A sNN matrix from graph-based clustering. For example,
the @graph$xxx_sNN matrix from Seurat object. The value can be
regarded as 'similarity' so 1-sNN can be regarded as the
'distance' measurement.}

\item{k}{Number of nearest neighbours used for smoothing. In
reality, top k+1 nearest neighbors will be considerred
including itself. Default: 10}

\item{method}{Either 'ave' or 'sum'. The smoothing result will be
either the average of the k-NNs or the sum of the k-NNs. Check
parameter 'strict' as well. Default: 'ave'}

\item{strict}{A boolean value. If TRUE, those nearest neighbors
with sNN=0 will be excluded, therefore, there may be less than
k NN considerred while calculating the average or sum; if
FALSE, all top nearest neighbors will be included. If 'method'
is 'ave', I recommend 'TRUE' ('FALSE' may be fine too);
however, if 'method' is 'sum', dropping sNN=0 neighbors may
make the cells/barcodes incomparable, so 'FALSE' is recommended
in this case. Default: TRUE}
}
\value{
A smoothed numeric matrix.
}
\description{
Note: sparse matrix will be converted to regular matrix, so memory
usage will be high.
}
