% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_getters.R
\name{last_build_errors}
\alias{last_build_errors}
\title{Retrieve latest maestro build errors}
\usage{
last_build_errors()
}
\value{
error messages
}
\description{
Gets the latest schedule build errors following use of \code{build_schedule()}. If
the build succeeded or \code{build_schedule()} has not been run it will be \code{NULL}.
}
\examples{
pipeline_dir <- tempdir()
create_pipeline("my_new_pipeline", pipeline_dir, open = FALSE, overwrite = TRUE)
build_schedule(pipeline_dir = pipeline_dir)
last_build_errors()
}
