% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_getters.R
\name{last_parsing_errors}
\alias{last_parsing_errors}
\title{Retrieve latest maestro parsing errors}
\usage{
last_parsing_errors()
}
\value{
error and trace output
}
\description{
Retrieve latest maestro parsing errors
}
