% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_getters.R
\name{last_run_errors}
\alias{last_run_errors}
\title{Retrieve latest maestro pipeline errors}
\usage{
last_run_errors()
}
\value{
error messages
}
\description{
Gets the latest pipeline errors following use of \code{run_schedule()}. If
the all runs succeeded or \code{run_schedule()} has not been run it will be \code{NULL}.
}
\examples{

last_run_errors()
}
