% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_getters.R
\name{last_runtime_errors}
\alias{last_runtime_errors}
\title{Retrieve latest maestro runtime errors}
\usage{
last_runtime_errors()
}
\value{
error message
}
\description{
Gets the latest runtime errors following use of `run_schedule()`
}
