% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_getters.R
\name{last_runtime_messages}
\alias{last_runtime_messages}
\title{Retrieve latest maestro runtime messages}
\usage{
last_runtime_messages()
}
\value{
messages
}
\description{
Gets the latest runtime messages following use of \code{run_schedule()}
}
