% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_getters.R
\name{last_runtime_warnings}
\alias{last_runtime_warnings}
\title{Retrieve latest maestro runtime warnings}
\usage{
last_runtime_warnings()
}
\value{
warning messages
}
\description{
Gets the latest runtime warnings following use of `run_schedule()`
}
