% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_errors.R
\name{latest_parsing_errors}
\alias{latest_parsing_errors}
\title{Retrieve latest baton parsing errors from}
\usage{
latest_parsing_errors()
}
\value{
error and trace output
}
\description{
Retrieve latest baton parsing errors from
}
