% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_errors.R
\name{latest_runtime_errors}
\alias{latest_runtime_errors}
\title{Retrieve latest maestro runtime errors}
\usage{
latest_runtime_errors()
}
\value{
error message
}
\description{
Retrieve latest maestro runtime errors
}
