% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maestro.R
\docType{package}
\name{maestro}
\alias{maestro-package}
\alias{maestro}
\title{\code{maestro} package}
\description{
Lightweight pipeline orchestration in R
}
\details{
Documentation: \href{https://github.com/whipson/maestro}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whipson/maestro}
  \item \url{https://whipson.github.io/maestro/}
  \item Report bugs at \url{https://github.com/whipson/maestro/issues}
}

}
\author{
\strong{Maintainer}: Will Hipson \email{will.e.hipson@gmail.com} (\href{https://orcid.org/0000-0002-3931-2189}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ryan Garnett \email{ryangarnett78@gmail.com} [contributor, copyright holder]
}

}
