% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cli.R
\name{maestro_dependency_graph_cli}
\alias{maestro_dependency_graph_cli}
\title{cli output for dependency tree}
\usage{
maestro_dependency_graph_cli(pipe_data, adjacency_list)
}
\arguments{
\item{pipe_data}{data.frame containing the pipe ids, names, and scripts (a private attribute of MaestroPipeline)}

\item{adjacency_list}{data.frame containing the from and to connections using the pipe_ids (a private attribute of MaestroPipeline)}
}
\value{
cli output
}
\description{
cli output for dependency tree
}
\keyword{internal}
