% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaestroPipelineDag.R
\name{MaestroPipelineDag}
\alias{MaestroPipelineDag}
\title{Special type of MaestroPipeline allowing for DAG pipelines
A MaestroPipelineDag is for multiple pipelines chained together}
\description{
Special type of MaestroPipeline allowing for DAG pipelines
A MaestroPipelineDag is for multiple pipelines chained together
}
\keyword{internal}
\section{Super class}{
\code{maestro::MaestroPipeline} -> \code{MaestroPipelineDag}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{adjacency_list}}{adjacency list data.frame defining the ordering of the pipelines}

\item{\code{pipeline_ids}}{data.frame linking a pipeline name to an integer id}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MaestroPipelineDag-run}{\code{MaestroPipelineDag$run()}}
\item \href{#method-MaestroPipelineDag-clone}{\code{MaestroPipelineDag$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="maestro" data-topic="MaestroPipeline" data-id="check_timeliness"><a href='../../maestro/html/MaestroPipeline.html#method-MaestroPipeline-check_timeliness'><code>maestro::MaestroPipeline$check_timeliness()</code></a></span></li>
<li><span class="pkg-link" data-pkg="maestro" data-topic="MaestroPipeline" data-id="get_artifacts"><a href='../../maestro/html/MaestroPipeline.html#method-MaestroPipeline-get_artifacts'><code>maestro::MaestroPipeline$get_artifacts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="maestro" data-topic="MaestroPipeline" data-id="get_errors"><a href='../../maestro/html/MaestroPipeline.html#method-MaestroPipeline-get_errors'><code>maestro::MaestroPipeline$get_errors()</code></a></span></li>
<li><span class="pkg-link" data-pkg="maestro" data-topic="MaestroPipeline" data-id="get_messages"><a href='../../maestro/html/MaestroPipeline.html#method-MaestroPipeline-get_messages'><code>maestro::MaestroPipeline$get_messages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="maestro" data-topic="MaestroPipeline" data-id="get_pipe_name"><a href='../../maestro/html/MaestroPipeline.html#method-MaestroPipeline-get_pipe_name'><code>maestro::MaestroPipeline$get_pipe_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="maestro" data-topic="MaestroPipeline" data-id="get_schedule"><a href='../../maestro/html/MaestroPipeline.html#method-MaestroPipeline-get_schedule'><code>maestro::MaestroPipeline$get_schedule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="maestro" data-topic="MaestroPipeline" data-id="get_status"><a href='../../maestro/html/MaestroPipeline.html#method-MaestroPipeline-get_status'><code>maestro::MaestroPipeline$get_status()</code></a></span></li>
<li><span class="pkg-link" data-pkg="maestro" data-topic="MaestroPipeline" data-id="get_warnings"><a href='../../maestro/html/MaestroPipeline.html#method-MaestroPipeline-get_warnings'><code>maestro::MaestroPipeline$get_warnings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="maestro" data-topic="MaestroPipeline" data-id="initialize"><a href='../../maestro/html/MaestroPipeline.html#method-MaestroPipeline-initialize'><code>maestro::MaestroPipeline$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="maestro" data-topic="MaestroPipeline" data-id="print"><a href='../../maestro/html/MaestroPipeline.html#method-MaestroPipeline-print'><code>maestro::MaestroPipeline$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaestroPipelineDag-run"></a>}}
\if{latex}{\out{\hypertarget{method-MaestroPipelineDag-run}{}}}
\subsection{Method \code{run()}}{
Runs the pipeline
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaestroPipelineDag$run(
  resources = list(),
  log_file = tempfile(),
  quiet = FALSE,
  log_file_max_bytes = 1e+06,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resources}}{named list of arguments and values to pass to the pipeline}

\item{\code{log_file}}{path to the log file for logging}

\item{\code{quiet}}{whether to silence console output}

\item{\code{log_file_max_bytes}}{maximum bytes of the log file before trimming}

\item{\code{...}}{additional arguments (unused)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaestroPipelineDag-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MaestroPipelineDag-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaestroPipelineDag$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
