% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_maestro_project.R
\name{new_maestro_project}
\alias{new_maestro_project}
\title{Creates a new maestro project}
\usage{
new_maestro_project(extension)
}
\arguments{
\item{extension}{file extension for the orchestrator (supports R, Quarto, and RMarkdown)}
}
\value{
invisible
}
\description{
Creates a new maestro project
}
