% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_rounding_unit}
\alias{parse_rounding_unit}
\title{Parse a time string}
\usage{
parse_rounding_unit(time_string)
}
\arguments{
\item{time_string}{string like 1 day, daily, 2 weeks, 12 hours, etc.}
}
\value{
nunit list
}
\description{
Parse a time string
}
\keyword{internal}
