% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_schedule.R
\name{plot_schedule}
\alias{plot_schedule}
\title{Plot pipeline runs from a schedule}
\usage{
plot_schedule(
  schedule,
  orch_frequency = "1 hour",
  start_datetime = lubridate::now(tzone = "UTC"),
  end_datetime = start_datetime + lubridate::days(7),
  fill = "#9012C790",
  interactive = TRUE
)
}
\arguments{
\item{schedule}{a table of scheduled pipelines generated from \code{build_schedule()}}

\item{orch_frequency}{of the orchestrator, a single string formatted like "1 day" or "2 weeks"}

\item{start_datetime}{date or timestamp for the start of the schedule}

\item{end_datetime}{date or timestamp for the end of the schedule}

\item{fill}{named or hex color for the tiles}

\item{interactive}{whether to generate an interactive plot using \code{plotly} (default is \code{TRUE})}
}
\value{
interactive (via plotly) or static plot (via ggplot)
}
\description{
Generates a plot of expected pipeline runs based on
a schedule and an orchestrator frequency.
}
\examples{
pipeline_dir <- tempdir()
create_pipeline("my_new_pipeline", pipeline_dir, open = FALSE, overwrite = TRUE)
schedule <- build_schedule(pipeline_dir = pipeline_dir)
plot_schedule(schedule, orch_frequency = "1 hour", interactive = FALSE)
}
