% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAAnnotateCelltype.R
\docType{methods}
\name{RNAAnnotateCelltype}
\alias{RNAAnnotateCelltype}
\title{Annotate scRNA-seq celltype based on marker genes}
\usage{
RNAAnnotateCelltype(
  RNA,
  genes,
  signatures = "human.immune.CIBERSORT",
  min.score = 0,
  orig.ident = NULL
)
}
\arguments{
\item{RNA}{Seurat object of clustered scRNA-seq dataset, generated by \code{\link{RNARunSeurat}} function.}

\item{genes}{Data frame of differential expressed genes, generated by \code{\link{RNARunSeurat}} function.}

\item{signatures}{The \code{signatures} can be character or data frame. If \code{signatures} is character, 
it should be "human.immune.CIBERSORT", "mouce.brain.ALLEN", "mouse.all.facs.TabulaMuris" or "mouse.all.droplet.TabulaMuris".
If \code{signatures} is a data frame, is should be a data frame of celltype signatures, with first column is celltype, and second column is
gene symbol. Check the immune signatures from CIBERSORT (Newman et al., Nature Method, 2015) in example for details. Default is "human.immune.CIBERSORT".}

\item{min.score}{Minimum score required. For one cluster, if the score for all celltypes are less than 
\code{min.score}, the cluster will be annotated as "Others". Default is 0.}

\item{orig.ident}{Orignal identity. If given, will use the original identity to annotate the celltypes. Default is NULL.}
}
\value{
A RNA Seurat object with annotated celltypes.
}
\description{
Annotate the celltype for all clusters based on the marker genes and the input celltype signatures for scRNA-seq dataset.
}
\examples{
data(pbmc.RNA)
data(human.immune.CIBERSORT)
pbmc.RNA.res <- RNARunSeurat(inputMat = pbmc.RNA, project = "PBMC.scRNA.Seurat", min.c = 10, min.g = 100, dims.use = 1:15)
pbmc.RNA.res$RNA <- RNAAnnotateCelltype(pbmc.RNA.res$RNA, pbmc.RNA.res$genes, human.immune.CIBERSORT, min.score = 0.05)
str(pbmc.RNA.res$RNA)

}
\author{
Chenfei Wang
}
