% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAAnnotateTranscriptionFactor.R
\docType{methods}
\name{RNAAnnotateTranscriptionFactor}
\alias{RNAAnnotateTranscriptionFactor}
\title{Predict driver transcrition factors for scRNA-seq clusters}
\usage{
RNAAnnotateTranscriptionFactor(
  RNA,
  genes,
  cluster = NULL,
  project = RNA@project.name,
  method = "LISA",
  rabit.path,
  organism = "GRCh38",
  top.tf = 10
)
}
\arguments{
\item{RNA}{Seurat object generated by \code{\link{RNARunSeurat}} function.}

\item{genes}{Data frame of differential expressed genes, generated by \code{\link{RNARunSeurat}}, \code{\link{FindMarkersMAESTRO}} or \code{\link{FindAllMarkersMAESTRO}} function.}

\item{cluster}{The cluster of given differential genes. 
For example, if the diff-genes are genrated using cluster 1,2,3 vs cluster 4,5,6, 
the cluster should be set as c(1,2,3,4,5,6).}

\item{project}{Output project name for the transcription factor analysis result. Default is the project name of the Seurat object.}

\item{method}{Use LISA or Rabit to identify the driver regulators. If using LISA, users will need to download and install annotation files. If Rabit is selected, should provide rabit annotation at the same time. Default is LISA.}

\item{rabit.path}{Path of the rabit annotation files. The rabit annoation for CistromeDB is available at https://github.com/liulab-dfci/MAESTRO.
Download and unzip the annotation, supply the path of the annotations to \code{rabit.path}.}

\item{lisa.mode}{Mode to Run LISA, "multi" or "one-vs-rest". Currently disbabled since we only support multi mode. }

\item{organism}{Organism for the dataset. Only support "GRCh38" and "GRCm38". Default is "GRCh38".}

\item{top.tf}{Output the target genes for the top TFs. Default is 10.}
}
\value{
A list of top enriched transcription factor family for each cluster. For each listed TF family,
TFs are ranked according to expression.
}
\description{
Predict driver transcrition factors for scRNA-seq clusters based on TF ChIP-seq peaks from CistromeDB.
}
\examples{
data(pbmc.RNA)
data(human.immune.CIBERSORT)
pbmc.RNA.res <- RNARunSeurat(inputMat = pbmc.RNA, project = "PBMC.scRNA.Seurat", min.c = 10, min.g = 100, dims.use = 1:15)
pbmc.RNA.res$RNA <- RNAAnnotateCelltype(pbmc.RNA.res$RNA, pbmc.RNA.res$genes, human.immune.CIBERSORT, min.score = 0.05)
pbmc.tfs <- RNAAnnotateTranscriptionFactor(pbmc.RNA.res$RNA, pbmc.RNA.res$genes, project = "PBMC.scRNA.TF", organism = "GRCh38", top.tf =10)
pbmc.tfs
}
\author{
Dongqing Sun, Chenfei Wang
}