% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNABatchCorrect.R
\docType{methods}
\name{RNABatchCorrect}
\alias{RNABatchCorrect}
\title{Remove batch effect in scRNA-seq dataset}
\usage{
RNABatchCorrect(
  RNA,
  batch,
  nfeatures = 2000,
  dims.use = 1:15,
  cluster.res = 0.6,
  only.pos = FALSE,
  genes.test.use = "presto",
  genes.cutoff = 1e-05,
  genes.pct = 0.1,
  genes.logfc = 0.25,
  runpca.agrs = list(),
  findneighbors.args = list(),
  findclusters.args = list(),
  ...
)
}
\arguments{
\item{RNA}{Seurat object of clustered scRNA-seq dataset, generated by \code{\link{RNARunSeurat}} function.}

\item{batch}{List of batch information, should be the same with the column names (cell name) with RNA object.}

\item{nfeatures}{Number of the features used in the CCA analysis, default is 2000, use top 2000 HVG features from each batch.}

\item{dims.use}{Number of dimensions used for UMAP analysis. Default is 1:15, use the first 15 PCs.}

\item{cluster.res}{Value of the clustering resolution parameter. Default is 0.6.}

\item{project}{Output project name. Default is "MAESTRO.coembedded".}
}
\value{
A list contains a RNA Seurat object with batch removed and a data frame for the cluster specific genes.
}
\description{
Remove batch effect from scRNA-seq dataset, batch information should be provided, remove the batch and perform clustering again.
}
\examples{

}
\author{
Chenfei Wang
}
