% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNACorrectCelltype.R
\docType{methods}
\name{RNACorrectCelltype}
\alias{RNACorrectCelltype}
\title{Correct scRNA-seq celltypes based on specific marker genes}
\usage{
RNACorrectCelltype(
  RNA,
  genes,
  celltype = "nonimmune",
  signatures,
  score.cutoff = NULL
)
}
\arguments{
\item{RNA}{Seurat object of clustered scRNA-seq dataset, generated by \code{\link{RNARunSeurat}} function.}

\item{genes}{Data frame of differential expressed genes, generated by \code{\link{RNARunSeurat}} function.}

\item{celltype}{Cell-type to be corrected. Users can choose from 'nonimmune', 'CD4T/CD8T', 'CD8Tex', 'Tprolif', 'NK', 'pDC'.
If non-immune cells need to be corrected, users should provide their own signatures.}

\item{signatures}{The \code{signatures} should be a data frame of cell-type signatures, with first column is cell-type, and second column is gene symbol.}

\item{score.cutoff}{Cutoff for cell-type signature score.}
}
\value{
A RNA Seurat object with corrected celltypes stored in \code{assign.MAESTRO} of \code{meta.data}.
}
\description{
Correct the automatic celltype annotation result for some clusters based on specific marker genes
}
\author{
Dongqing Sun
}
