% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNACountToFPKM.R
\docType{methods}
\name{RNACountToFPKM}
\alias{RNACountToFPKM}
\title{Convert scRNA-seq count matrix to FPKM matrix}
\usage{
RNACountToFPKM(countMat, idType = "Ensembl", organism = "GRCh38")
}
\arguments{
\item{countMat}{Input count matrix, with genes as rows and cells as columns.}

\item{idType}{Type of gene name, available options are "Ensembl" or "Symbol". Default is "Symbol".}

\item{organism}{Ogranism used, currently only support human "GRCh38" and mouse "GRCm38". Default is "GRCh38".}
}
\value{
A matrix of FPKM, with genes as rows and cells as columns.
}
\description{
Convert scRNA-seq count matrix to FPKM matrix for downstream analysis.
}
\examples{
data(pbmc.RNA)
pbmc.RNA.FPKM <- RNACountToFPKM(pbmc.RNA, idType = "Symbol", organism = "GRCh38")
head(pbmc.RNA.FPKM[,1:10])

}
\author{
Chenfei Wang, Dongqing Sun
}
