% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNACountToTPM.R
\docType{methods}
\name{RNACountToTPM}
\alias{RNACountToTPM}
\title{Convert scRNA-seq count matrix to TPM matrix}
\usage{
RNACountToTPM(countMat, idType = "Ensembl", organism = "GRCh38")
}
\arguments{
\item{countMat}{Input count matrix, with genes as rows and cells as columns.}

\item{idType}{Type of gene name, available options are "Ensembl" or "Symbol". Default is "Ensembl".}

\item{organism}{Ogranism used, currently only support human "GRCh38" and mouse "GRCm38". Default is "GRCh38".}
}
\value{
A matrix of TPM, with genes as rows and cells as columns.
}
\description{
Convert scRNA-seq count matrix to TPM matrix for downstream analysis.
}
\examples{
data(pbmc.RNA)
pbmc.RNA.TPM <- RNACountToTPM(pbmc.RNA, idType = "Ensembl", organism = "GRCh38")
head(pbmc.RNA.TPM[,1:10])

}
\author{
Chenfei Wang, Dongqing Sun
}
