% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNADiffusionMap.R
\docType{methods}
\name{RNADiffusionMap}
\alias{RNADiffusionMap}
\title{Perform lineage tracing analysis in scRNA-seq dataset}
\usage{
RNADiffusionMap(
  RNA,
  project = RNA@project.name,
  start.genes,
  end.genes,
  clusters = unique(RNA@meta.data$seurat_clusters)
)
}
\arguments{
\item{RNA}{Seurat object of clustered scRNA-seq dataset, generated by \code{\link{RNARunSeurat}} function.}

\item{project}{Output project name. Default is the RNA Seurat object name.}

\item{start.genes}{Genes used for the start.}

\item{end.genes}{Genes used for the end.}

\item{clusters}{Clusters for perform diffusion map analysis, default use all the clusters.}
}
\value{
A diffusion map object with pseudo time information.
}
\description{
Perform lineage tracing analysis in scRNA-seq dataset using Diffusion map.
}
\examples{

}
\author{
Chenfei Wang
}
