% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAEnsemblToSymbol.R
\docType{methods}
\name{RNAEnsemblToSymbol}
\alias{RNAEnsemblToSymbol}
\title{Convert Ensembl ID to Gene Symbol}
\usage{
RNAEnsemblToSymbol(countMat, organism = "GRCh38")
}
\arguments{
\item{countMat}{Input count matrix, with genes as rows and cells as columns.}

\item{organism}{Ogranism used, currently only support human "GRCh38" and mouse "GRCm38". Default is "GRCh38".}
}
\value{
A matrix of count, with genes as rows and cells as columns.
}
\description{
Convert Ensembl ID to Gene Symbol for count matrix.
}
\examples{
data(pbmc.RNA)
pbmc.RNA.symbol <- RNAEnsemblToSymbol(pbmc.RNA, organism = "GRCh38")
head(pbmc.RNA.symbol)

}
\author{
Chenfei Wang, Dongqing Sun
}
