% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAFilteringPlot.R
\docType{methods}
\name{RNAFilteringPlot}
\alias{RNAFilteringPlot}
\title{Cell filtering plot for scRNA-seq analysis}
\usage{
RNAFilteringPlot(filepath, UMI.cutoff = 500, gene.number.cutoff = 200,
  name)
}
\arguments{
\item{filepath}{Path of the scRNA-seq mapping sumary file generated by \code{MAESTRO}.}

\item{UMI.cutoff}{UMI cutoff. Cells with less than \code{UMI.cutoff} UMIs will be classified as non-cells.
Default is 500.}

\item{gene.number.cutoff}{Gene number cutoff. Cells covered less than \code{gene.number.cutoff} will be 
classified as non-cells. Default is 200.}

\item{name}{Name for the output cell filtering plot and valid cells file.}
}
\description{
Filtering the cells and non-cells based on number of UMIs and covered genes in each cell.
}
\author{
Dongqing Sun, Chenfei Wang
}
