% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAGCcontentPlot.R
\docType{methods}
\name{RNAGCcontentPlot}
\alias{RNAGCcontentPlot}
\title{GC content plot for scRNA-seq QC}
\usage{
RNAGCcontentPlot(filepath, name)
}
\arguments{
\item{filepath}{Path of the GC content file generated by \code{MAESTRO}.}

\item{name}{Name for the output GC content plot.}
}
\description{
Generate the GC content plot for scRNA-seq quality control analysis.
}
\author{
Dongqing Sun, Chenfei Wang
}
