% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAGenebodyCoveragePlot.R
\docType{methods}
\name{RNAGenebodyCoveragePlot}
\alias{RNAGenebodyCoveragePlot}
\title{Genebody coverage plot for scRNA-seq analysis}
\usage{
RNAGenebodyCoveragePlot(filepath, name)
}
\arguments{
\item{filepath}{Path of the genebody coverage file generated by \code{MAESTRO}.}

\item{name}{Name for the output genebody coverage plot.}
}
\description{
Generate the genebody coverage plot for scRNA-seq quality control analysis.
}
\author{
Dongqing Sun, Chenfei Wang
}
