% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNANucleotidePlot.R
\docType{methods}
\name{RNANucleotidePlot}
\alias{RNANucleotidePlot}
\title{NVC plot for scRNA-seq QC}
\usage{
RNANucleotidePlot(filepath, name)
}
\arguments{
\item{filepath}{Path of the nucleotide composition file generated by \code{MAESTRO}.}

\item{name}{Name for the output NVC plot.}
}
\description{
Generate the NVC(Nucleotide versus cycle) plot for scRNA-seq quality control analysis.
}
\author{
Dongqing Sun, Chenfei Wang
}
