% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAReadDistrPlot.R
\docType{methods}
\name{RNAReadDistrPlot}
\alias{RNAReadDistrPlot}
\title{Read distribution plot for scRNA-seq analysis}
\usage{
RNAReadDistrPlot(bamstat.filepath, readdistr.filepath, name, platform)
}
\arguments{
\item{bamstat.filepath}{Path of the scRNA-seq mapping summary file generated by \code{MAESTRO}. 
For 10x-genomics, the file is metrics_summary.csv. For other platform, the file is}

\item{readdistr.filepath}{Path of the scRNA-seq read distribution file generated by \code{MAESTRO}.}

\item{name}{Name for the output read distribution plot.}

\item{platform}{Denote platform of the scRNA-seq data, available options are "Smartseq2",
"10x-genomics","Dropseq". Default platform is "10x-genomics".}
}
\description{
Summary plot of total reads, duplicate reads ratio, mappability, exon and intron reads ratio for scRNA-seq analysis.
}
\author{
Dongqing Sun, Chenfei Wang
}
