% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAReadQualityPlot.R
\docType{methods}
\name{RNAReadQualityPlot}
\alias{RNAReadQualityPlot}
\title{Read quality plot for scRNA-seq analysis}
\usage{
RNAReadQualityPlot(filepath, name)
}
\arguments{
\item{filepath}{Path of the read quality r file generated by \code{MAESTRO}.}

\item{name}{Name for the output read quality plot.}
}
\description{
Generate the read quality plot for scRNA-seq quality control analysis.
}
\author{
Dongqing Sun, Chenfei Wang
}
