% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNARunPagoda.R
\docType{methods}
\name{RNARunPagoda}
\alias{RNARunPagoda}
\title{Clustering analysis for scRNA-seq data using Pagoda2}
\usage{
RNARunPagoda(
  inputMat,
  project = "MAESTRO.scRNA.Pagoda",
  min.c = 10,
  min.g = 200,
  nPCs = 100,
  nKs = 30
)
}
\arguments{
\item{inputMat}{Input unnormalized matrix, with genes as rows and cells as columns. Could be count matrix or TPM, FPKM matrix.}

\item{project}{Output project name. Default is "MAESTRO.scRNA.Pagoda".}

\item{min.c}{Minimum number of cells required for a gene. Will exclude the genes from input matrix if they only expressed in 
less than \code{min.c} cells. Default is 10.}

\item{min.g}{Minimum number of genes required for a cell. Will exclude the cells from input matrix if less than \code{min.g}
genes are deteced in one cell. Default is 200.}

\item{nPCs}{Number of PCs considered in the clustering analysis. Default is 100.}

\item{nKs}{Number of K for the KNN graph. Default is 30.}
}
\value{
A list contains a Pagoda2 object and a list for the cluster specific genes.
}
\description{
Clustering analysis for scRNA-seq dataset using Pagoda2 (version 0.1.0). Including feature selection, dimension reduction, clustering and tSNE visualization. To run this function, you must first install Pagoda2 (see from https://github.com/hms-dbmi/pagoda2).
}
\examples{
data(pbmc.RNA)
pbmc <- RNARunPagoda(inputMat = pbmc.RNA, project = "PBMC.scRNA.Pagoda2")
pbmc$RNA
head(pbmc$genes)

}
\author{
Chenfei Wang
}
