% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNARunSeurat.R
\docType{methods}
\name{RNARunSeurat}
\alias{RNARunSeurat}
\title{Clustering analysis for scRNA-seq data using Seurat}
\usage{
RNARunSeurat(
  inputMat,
  type = "matrix",
  project = "MAESTRO.scRNA.Seurat",
  orig.ident = NULL,
  min.c = 10,
  min.g = 200,
  mito = FALSE,
  variable.genes = 2000,
  organism = "GRCh38",
  dims.use = 1:15,
  cluster.res = 0.6,
  only.pos = FALSE,
  genes.test.use = "presto",
  genes.cutoff = 1e-05,
  genes.pct = 0.1,
  genes.logfc = 0.25,
  runpca.agrs = list(),
  findneighbors.args = list(),
  findclusters.args = list(),
  ...
)
}
\arguments{
\item{inputMat}{Input unnormalized matrix, with genes as rows and cells as columns. Could be count matrix or TPM, FPKM matrix.}

\item{type}{Type of the input matrix. Default is "matrix". Set to "object" if the input is Seurat object.}

\item{project}{Output project name. Default is "MAESTRO.scRNA.Seurat".}

\item{orig.ident}{Orginal identity for the input cells. If supplied, should keep the same order with the column name of the gene x cell matrix.}

\item{min.c}{Minimum number of cells required for a gene. Will exclude the genes from input matrix if they only expressed in 
less than \code{min.c} cells. Default is 10. See \code{\link{CreateSeuratObject}} for details.}

\item{min.g}{Minimum number of genes required for a cell. Will exclude the cells from input matrix if less than \code{min.g}
genes are deteced in one cell. Default is 200. See \code{\link{CreateSeuratObject}} for details.}

\item{mito}{Whether or not to check and normalized accroding to the fraction of mitochondria reads and ercc spike-in reads.
If \code{mito} is True, those cells with mitochondria reads or ercc spike-in reads higher than 0.05 will be filtered.
Default is FALSE.}

\item{variable.genes}{Number of variable genes considered in the clustering analysis. Default is 2000. See \code{\link{FindVariableFeatures}} for details.}

\item{organism}{Organism for the dataset. Only support "GRCh38" and "GRCm38". Default is "GRCh38".}

\item{dims.use}{Number of dimensions used for UMAP analysis. Default is 1:15, use the first 15 PCs.}

\item{cluster.res}{Value of the clustering resolution parameter. Default is 0.6.}

\item{only.pos}{If seting true, only positive genes will be output. Default is False.}

\item{genes.test.use}{Denotes which test to use to identify genes. Default is "presto", a fast version of Wilcoxon Rank Sum test. See \code{\link{FindAllMarkersMAESTRO}} for details.}

\item{genes.cutoff}{Identify differential expressed genes with adjusted p.value less than \code{genes.cutoff} as cluster speficic genes}

\item{genes.pct}{Only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently detected Default is 0.1}

\item{genes.logfc}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.2 Increasing logfc.threshold speeds up the function, but can miss weaker signals.
for each cluster. Default cutoff is 1E-5.}

\item{findneighbors.args}{Extra arguments passed to \code{\link{FindNeighbors}}.}

\item{findclusters.args}{Extra arguments passed to \code{\link{FindClusters}}.}

\item{\dots}{Extra arguments passed to \code{\link{RunUMAP}}.}

\item{runpca.args}{Extra arguments passed to \code{\link{RunPCA}}.}
}
\value{
A list contains a RNA Seurat object and a data frame for the cluster specific genes.
}
\description{
Clustering analysis for scRNA-seq dataset using Seurat(version >=3.0.1). Including normalization, feature selection, dimension reduction, clustering and UMAP visualization. To run UMAP analysis, you must first install the umap-learn python package (e.g. via \code{pip install umap-learn}).
}
\examples{
data(pbmc.RNA)
pbmc.RNA.res <- RNARunSeurat(inputMat = pbmc.RNA, project = "PBMC.scRNA.Seurat", min.c = 10, min.g = 100, dims.use = 1:15)
str(pbmc.RNA.res$RNA)
head(pbmc.RNA.res$genes)

}
\author{
Chenfei Wang
}
