% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_time}
\alias{round_time}
\title{Round time formatted}
\usage{
round_time(check_datetime, unit_value)
}
\arguments{
\item{check_datetime}{POSIXct object}

\item{unit_value}{unit to round to}
}
\value{
character
}
\description{
Round time formatted
}
