% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_entry_from_script.R
\name{schedule_entry_from_script}
\alias{schedule_entry_from_script}
\title{Create schedule table entry from a script}
\usage{
schedule_entry_from_script(script_path)
}
\arguments{
\item{script_path}{path to script}
}
\value{
data.frame row
}
\description{
Create schedule table entry from a script
}
