% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_pipelines.R
\name{select_pipelines}
\alias{select_pipelines}
\title{Select pipelines}
\usage{
select_pipelines(.data, orch_interval, orch_unit, check_datetime = Sys.time())
}
\arguments{
\item{.data}{data.frame with pipeline name, interval, frequency, and start datetime}

\item{orch_interval}{a numeric value representing how often the orchestration runs}

\item{orch_unit}{unit of how often the orchestration runs (i.e. min, hour, day, month, etc.)}

\item{check_datetime}{datetime used to check against the pipeline start datetime, default is the system time}
}
\value{
data.frame
}
\description{
Select pipelines
}
