% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_baton_project.R
\name{setup_baton_project}
\alias{setup_baton_project}
\title{Setup a Baton Project Structure}
\usage{
setup_baton_project(
  as_quarto = TRUE,
  cached_table = FALSE,
  multithreaded = TRUE
)
}
\arguments{
\item{as_quarto}{should the orchestrator be a Quarto doc}

\item{cached_table}{use a cached schedule table or in memory}

\item{multithreaded}{run multithreaded}
}
\value{
quarto or R script
}
\description{
asdasdasdasdasdasd
}
\details{
asdasdasdasdasdasdas
}
