% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_maestro_project.R
\name{setup_maestro_project}
\alias{setup_maestro_project}
\title{Setup a maestro Project Structure}
\usage{
setup_maestro_project(
  as_quarto = TRUE,
  cached_table = FALSE,
  multithreaded = TRUE
)
}
\arguments{
\item{as_quarto}{should the orchestrator be a Quarto doc}

\item{cached_table}{use a cached schedule table or in memory}

\item{multithreaded}{run multithreaded}
}
\value{
quarto or R script
}
\description{
Setup a maestro Project Structure
}
