% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeTFenrichment.R
\docType{methods}
\name{VisualizeTFenrichment}
\alias{VisualizeTFenrichment}
\title{Scatter plot to show the enrichment of TFs 
Draw a scatter plot to show the enrichment of TFs in specific clusters, based on cluster-specific TFs or cluster-specific peaks. Currently only support LISA for scRNA-seq analysis and GIGGLE for scATAC-seq analysis.}
\usage{
VisualizeTFenrichment(
  TFs = NULL,
  cluster.1,
  cluster.2 = NULL,
  type = "RNA",
  SeuratObj,
  visual.topnumber = 10,
  visual.totalnumber = 100,
  LISA.table = NULL,
  GIGGLE.table = NULL,
  width = 5,
  height = 4.5,
  name = "TFEnrichmentPlot"
)
}
\arguments{
\item{TFs}{TFs to plot.  If no TFs is given, by default visualize the top \code{visual.topnumber} regulators in that cluster. For "Incorparated" result, is no TFs is given,
using the rank from GIGGLE analysis on scATAC-seq dataset.}

\item{cluster.1}{ClusterID for which the regulators need to be plotted.}

\item{cluster.2}{For integrated result, cluster.1 represent the clusterID for scRNA-seq, cluster.2 represent the clusterID for scATAC-seq, default set to NULL when only
RNA or ATAC is analyzed.}

\item{type}{Type of the dataset. If \code{type} is "RNA" or "Integrated", the expression level of TFs will be visualized. 
If \code{type} is "ATAC", the regulatory potential will be used for visualization. Default is "RNA".}

\item{SeuratObj}{A Seurat object. If \code{type} is "RNA"or "Integrated", \code{SeuratObj} should be the Seurat object generated by \code{\link{RNARunSeurat}} function.
If \code{type} is "ATAC", \code{SeuratObj} should be a Seurat object created using RP matrix achievesd from \code{\link{ATACAnnotateCelltype}} function.}

\item{visual.topnumber}{If "top" is used, select top ranked regulators for labeling, default is top 10.}

\item{visual.totalnumber}{Number of total visualized regulators, default is top 100.}

\item{LISA.table}{Path of the score summary tale for LISA, default is \code{project.name_}TF_lisa.txt.}

\item{GIGGLE.table}{Path of the score summary tale for GIGGLE, default is \code{project.name_}TF_giggle.txt.}

\item{width}{Width of the plots.}

\item{height}{Height of the plots.}

\item{name}{Output name of the plots. Default is "TFEnrichmentPlot".}
}
\value{
A ggplot object which can be visulaized by \code{link{print}}.
}
\description{
Scatter plot to show the enrichment of TFs 
Draw a scatter plot to show the enrichment of TFs in specific clusters, based on cluster-specific TFs or cluster-specific peaks. Currently only support LISA for scRNA-seq analysis and GIGGLE for scATAC-seq analysis.
}
\examples{
data(pbmc.RNA)
pbmc <- RNARunSeurat(inputMat = pbmc.RNA, project = "PBMC.scRNA.Seurat")
pbmc.tfs <- RNAAnnotateTranscriptionFactor(pbmc$RNA, pbmc$TFs, project = "PBMC.scRNA.TF", method = "LISA")

# TFs from MAESTRO result
tfs = sapply(pbmc.tfs[[1]], function(x){return(unlist(strsplit(x, split = " | ", fixed = TRUE))[1])})
VisualizeTFenrichment(TFs = tfs, cluster.1 = 0, type = "RNA", SeuratObj = pbmc$RNA, visual.topnumber = 10, visual.totalnumber = 100, width = 6, height = 3, name = "PBMC.scRNA.tf")

# TFs of interest
VisualizeTFenrichment(TFs = c("CEBPA","CEBPB","CEBPD"), cluster.1 = 0, type = "RNA", SeuratObj = pbmc$RNA, visual.totalnumber = 100, width = 6, height = 3, name = "PBMC.scRNA.tf")


}
\author{
Chenfei Wang, Dongqing Sun
}
