% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACRunSeurat.R
\docType{methods}
\name{ATACRunSeurat}
\alias{ATACRunSeurat}
\title{Clustering analysis for scATAC-seq data using Seurat}
\usage{
ATACRunSeurat(
  inputMat,
  project = "MAESTRO.scATAC.Seurat",
  orign.ident = NULL,
  method = "LSI",
  min.c = 50,
  min.p = 500,
  dims.use = 1:30,
  cluster.res = 0.6,
  peaks.test.use = "wilcox",
  peaks.cutoff = 1e-05
)
}
\arguments{
\item{inputMat}{Input unnormalized count matrix, with peaks as rows and cells as columns. Rownames should be like "chromosome_peakstart_peakend",
for example "chr10_100020591_100020841".}

\item{project}{Output project name. Default is "MAESTRO.scATAC.Seurat".}

\item{orign.ident}{Orginal identity for the input cells. If supplied, should keep the same order with the column name of the peak x cell matrix.}

\item{method}{Methods for dimension reduction, available options are LSI and PCA. Default is "LSI".}

\item{min.c}{Minimum number of cells required for a peak. Will exclude the peaks from input matrix if they only identified in 
less than \code{min.c} cells. Default is 50. See \code{link{CreateSeuratObject}} for details.}

\item{min.p}{Minimum number of peaks required for a cell. Will exclude the cells from input matrix if less than \code{min.p}
peaks are deteced in one cell. Default is 500. See \code{link{CreateSeuratObject}} for details.}

\item{dims.use}{Number of dimensions used for UMAP analysis. Default is 1:30, use the first 30 PCs.}

\item{cluster.res}{Value of the clustering resolution parameter. Default is 0.6.}

\item{peaks.test.use}{Denotes which test to use to identify differnetial peaks. Default is "wilcox". Available options are "bimod", "roc" and "t".}

\item{peaks.cutoff}{Identify differential peaks with adjusted p.value less than \code{peaks.cutoff} as cluster specific peaks
for each cluster. Default cutoff is 1E-5.}
}
\value{
A list contains a Seurat object for ATAC clusters and a data frame of cluster specific peaks.
}
\description{
Clustering analysis for scATAC-seq dataset using Seurat(version >=3.0.1). Including normalization, LSI/PCA dimension reduction, clustering and UMAP visualization. To run UMAP analysis, you must first install the umap-learn python package (e.g. via \code{pip install umap-learn}).
}
\examples{
data(pbmc.ATAC)
pbmc.ATAC.res <- ATACRunSeurat(inputMat = pbmc.ATAC, project = "PBMC.scATAC.Seurat", method = "LSI")
str(pbmc.ATAC.res$ATAC)
head(pbmc.ATAC.res$peaks)

}
\author{
Chenfei Wang
}
