% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_schedule.R
\name{build_schedule}
\alias{build_schedule}
\title{Build schedule table}
\usage{
build_schedule(pipeline_dir = "./pipelines")
}
\arguments{
\item{pipeline_dir}{path to directory containing the pipeline scripts}
}
\value{
data.frame
}
\description{
Build schedule table
}
