% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_schedule_entry.R
\name{build_schedule_entry}
\alias{build_schedule_entry}
\title{Create schedule table entry from a script}
\usage{
build_schedule_entry(script_path)
}
\arguments{
\item{script_path}{path to script}
}
\value{
data.frame row
}
\description{
Create schedule table entry from a script
}
