% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pipeline_next_schedule.R
\name{check_pipeline_next_schedule}
\alias{check_pipeline_next_schedule}
\title{Check pipeline next schedule}
\usage{
check_pipeline_next_schedule(
  orch_frequency,
  check_datetime,
  pipeline_seconds,
  pipeline_datetime
)
}
\arguments{
\item{orch_frequency}{frequency of the orchestrator, a single string formatted like \link{number} \link{units} (e.g., "1 day")}

\item{check_datetime}{datetime to be used to check if the pipeline should run}

\item{pipeline_seconds}{unit of how often the pipeline runs (i.e. min, hour, day, month, etc.)}

\item{pipeline_datetime}{datetime of the first time the pipeline is to run}
}
\value{
list
}
\description{
Checks if a pipeline is scheduled to run and when it is next supposed to run
}
