% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pipelines.R
\name{check_pipelines}
\alias{check_pipelines}
\title{Check which pipelines are scheduled to run and when next pipelines will run}
\usage{
check_pipelines(.data, orch_frequency, check_datetime)
}
\arguments{
\item{.data}{data.frame with pipeline name, frequency, and start datetime}

\item{orch_frequency}{frequency of the orchestrator, a single string formatted like \link{number} \link{units} (e.g., "1 day")}

\item{check_datetime}{datetime used to check against the pipeline start datetime}
}
\value{
list
}
\description{
Check which pipelines are scheduled to run and when next pipelines will run
}
