% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pipelines.R
\name{check_pipelines}
\alias{check_pipelines}
\title{Check which pipelines are scheduled to run and when next pipelines will run}
\usage{
check_pipelines(.data, orch_interval, orch_frequency, check_datetime)
}
\arguments{
\item{.data}{data.frame with pipeline name, interval, frequency, and start datetime}

\item{orch_interval}{a numeric value representing how often the orchestration runs}

\item{orch_frequency}{unit of how often the orchestration runs (i.e. min, hour, day, month, etc.)}

\item{check_datetime}{datetime used to check against the pipeline start datetime}
}
\value{
list
}
\description{
Check which pipelines are scheduled to run and when next pipelines will run
}
