% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_maestro.R
\name{create_maestro}
\alias{create_maestro}
\title{Creates a new maestro project}
\usage{
create_maestro(path = ".", extension, ...)
}
\arguments{
\item{path}{file path for the orchestrator script}

\item{extension}{file extension for the orchestrator (supports R, Quarto, and RMarkdown)}
}
\value{
invisible
}
\description{
Creates a new maestro project
}
