% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pipeline.R
\name{create_pipeline}
\alias{create_pipeline}
\title{Create a new pipeline in a pipelines directory}
\usage{
create_pipeline(
  pipe_name,
  pipeline_dir = "pipelines",
  frequency = NULL,
  interval = NULL,
  start_time = NULL,
  tz = NULL,
  open = interactive()
)
}
\arguments{
\item{pipe_name}{name of the pipeline and function}

\item{frequency}{how often the pipeline should run (e.g., hourly, daily, etc.). Fills in maestroFrequency tag}

\item{interval}{units of frequency between runs (e.g., 1, 2, 7). Fills in maestroFrequency tag}

\item{start_time}{start time of the pipeline schedule. Fills in maestroStartTime tag}

\item{tz}{timezone that pipeline will be scheduled in. Fills in maestroTz tag}

\item{open}{whether or not to open the script upon creation}
}
\value{
invisible
}
\description{
Create a new pipeline in a pipelines directory
}
