% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maestro_tags.R
\name{maestro_tags}
\alias{maestro_tags}
\title{Maestro Tags}
\description{
maestro tags are roxygen2 comments for configuring the scheduling and execution
of pipelines.
}
\details{
maestro tags follow the format \verb{#' @maestroTagName}
\subsection{Tag List}{
\itemize{
\item maestroFrequency: Time unit for scheduling (e.g., day, week, month, etc.)
\item maestroInterval: Number of time units between executions (e.g., 1, 2, 3)
\item maestroLogLevel: Threshold for logging when logging is requested by orchestrator (e.g., INFO, WARN, ERROR, etc.)
\item maestroSkip: Skips the pipeline when running orchestrator (presence of tag indicates skip)
\item maestroStartTime: yyyy-MM-dd HH:MM:SS for the start time of the pipeline
\item maestroTz: Timezone of the start time (e.g., UTC); see \code{OlsonNames()} for all timezones
}
}
}
