% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maestro_tags.R
\name{maestro_tags}
\alias{maestro_tags}
\title{Maestro Tags}
\description{
maestro tags are roxygen2 comments for configuring the scheduling and execution
of pipelines.
}
\details{
maestro tags follow the format \verb{#' @maestroTagName}
\subsection{Tag List}{\tabular{lllll}{
   tagName \tab description \tab value \tab examples (comma sep.) \tab default \cr
   maestroFrequency \tab Time unit for scheduling \tab string \tab 1 hour, 3 days, 5 weeks \tab 1 day \cr
   maestroLogLevel \tab Threshold for logging when logging is requested \tab string \tab INFO, WARN, ERROR \tab INFO \cr
   maestroSkip \tab Skips the pipeline when running (presence of tag indicates to skip) \tab n/a \tab  \tab  \cr
   maestroStartTime \tab Start time of the pipeline; sets the point in time for recurrence \tab date or timestamp \tab 1970-01-01 00:00:00, 2024-03-28 \tab 1970-01-01 00:00:00 \cr
   maestroTz \tab Timezone of the start time \tab string \tab UTC, America/Halifax \tab UTC \cr
}

}
}
