% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAAnnotateCelltypeCluster.R
\docType{methods}
\name{RNAAnnotateCelltypeCluster}
\alias{RNAAnnotateCelltypeCluster}
\title{Annotate celltype based on marker genes for one cluster}
\usage{
RNAAnnotateCelltypeCluster(
  genes,
  signatures = "human.immune.CIBERSORT",
  cluster = 0
)
}
\arguments{
\item{genes}{Data frame of differential expressed genes, generated by \code{\link{RNARunSeurat}} function.}

\item{signatures}{The \code{signatures} can be character or data frame. If \code{signatures} is character, 
it should be "human.immune.CIBERSORT", "mouce.brain.ALLEN", "mouse.all.facs.TabulaMuris" or "mouse.all.droplet.TabulaMuris".
If \code{signatures} is a data frame, is should be a data frame of celltype signatures, with first column is celltype, and second column is
gene symbol. Check the immune signatures from CIBERSORT (Newman et al., Nature Method, 2015) in example for details. Default is "human.immune.CIBERSORT".}

\item{cluster}{Cluster ID. Default is 0. Check the celltype identify of cluster 0.}
}
\value{
An list of ranked celltype score based on given signatures.
}
\description{
Annotate the celltype for one cluster based on the marker genes of that cluster and celltype signatures.
}
\examples{
data(pbmc.RNA)
data(human.immune.CIBERSORT)
pbmc.RNA.res <- RNARunSeurat(inputMat = pbmc.RNA, project = "PBMC.scRNA.Seurat", min.c = 10, min.g = 100, dims.use = 1:15)
celtype.score <- RNAAnnotateCelltypeCluster(pbmc.RNA.res$genes, human.immune.CIBERSORT, cluster = 0)
celtype.score

}
\author{
Chenfei Wang
}
