% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAAnnotateTranscriptionFactor.R
\docType{methods}
\name{RNAAnnotateTranscriptionFactor}
\alias{RNAAnnotateTranscriptionFactor}
\title{Predict driver transcrition factors for scRNA-seq clusters}
\usage{
RNAAnnotateTranscriptionFactor(RNA, genes, project, method = "RABIT",
  rabit.path, organism = "GRCh38", top.tf = 10)
}
\arguments{
\item{RNA}{Seurat object generated by \code{\link{RNARunSeurat}} function.}

\item{genes}{Data frame of differential expressed geness, generated by \code{\link{RNARunSeurat}} function.}

\item{project}{Output project name for the transcription factor analysis result. Default is the project name of the Seurat object.}

\item{method}{Use RABIT or LISA to identify the driver regulators. If Rabit is selected, should provide rabit annotation at the
same time. LISA do not need annotations for this step. Default is Rabit.}

\item{rabit.path}{Path of the rabit annotation files. The rabit annoation for CistromeDB is available at https://github.com/liulab-dfci/MAESTRO.
Download and unzip the annotation, supply the path of the annotations to \code{rabit.path}.}

\item{organism}{Organism for the dataset. Only support "GRCh38" and "GRCm38". Default is "GRCh38".}

\item{top.tf}{Output the target genes for the top TFs. Default is 10.}
}
\value{
A list of top enriched transcription factor family for each cluster. For each listed TF family,
TFs are ranked according to expression.
}
\description{
Predict driver transcrition factors for scRNA-seq clusters based on TF ChIP-seq peaks from CistromeDB. To run this function, you must first install the rabit package (see from http://rabit.dfci.harvard.edu/).
}
\examples{
data(pbmc.RNA)
data(human.immune.CIBERSORT)
pbmc.RNA.res <- RNARunSeurat(inputMat = pbmc.RNA, project = "PBMC.scRNA.Seurat", min.c = 10, min.g = 100, dims.use = 1:15)
pbmc.RNA.res$RNA <- RNAAnnotateCelltype(pbmc.RNA.res$RNA, pbmc.RNA.res$genes, human.immune.CIBERSORT, min.score = 0.05)
pbmc.tfs <- RNAAnnotateTranscriptionFactor(pbmc.RNA.res$RNA, pbmc.RNA.res$genes, project = "PBMC.scRNA.TF", rabit.path = "/homes/cwang/annotations/rabit")
pbmc.tfs

}
\author{
Dongqing Sun, Chenfei Wang
}
