% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_schedule.R
\name{run_schedule}
\alias{run_schedule}
\title{Run a schedule}
\usage{
run_schedule(
  schedule,
  orch_interval = 1,
  orch_unit = "day",
  check_datetime = lubridate::now(tzone = "UTC"),
  resources = list(),
  run_all = FALSE,
  n_show_not_run = 5,
  quiet = FALSE
)
}
\arguments{
\item{schedule}{a table of scheduled pipelines generated from `build_schedule()`}

\item{orch_interval}{a numeric value representing how often the orchestration runs}

\item{orch_unit}{unit of how often the orchestration runs (i.e. min, hour, day, month, etc.)}

\item{check_datetime}{datetime used to check against the pipeline start datetime}

\item{resources}{list of shared resources made available to pipelines as needed}

\item{run_all}{run all pipelines regardless of the schedule (default is `FALSE`) - useful for testing}

\item{n_show_not_run}{show number of pipelines that did not run and when they will run next.}

\item{quiet}{print metrics to the console (default = `TRUE`)}
}
\value{
invisible
}
\description{
Given a schedule in a `maestro` project, runs the pipelines that are scheduled to execute
based on the current time.
}
