% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_schedule.R
\name{run_schedule}
\alias{run_schedule}
\title{Run a schedule}
\usage{
run_schedule(schedule, resources = list())
}
\arguments{
\item{schedule}{a table of scheduled pipelines generated from `build_schedule()`}

\item{resources}{list of shared resources made available to pipelines as needed}
}
\value{
invisible
}
\description{
Run a schedule
}
