% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_schedule_entry.R
\name{run_schedule_entry}
\alias{run_schedule_entry}
\title{Runs a single pipeline}
\usage{
run_schedule_entry(
  script_path,
  pipe_name,
  resources = list(),
  log_file = NULL,
  log_level = "INFO",
  log_file_max_bytes
)
}
\arguments{
\item{script_path}{path to the script containing the pipeline}

\item{pipe_name}{name of the pipeline}

\item{resources}{list of resources for the pipeline}

\item{log_file}{path to log file}

\item{log_level}{log level}

\item{log_file_max_bytes}{numeric specifying the maximum number of bytes allowed in the log file before purging the log (within a margin of error)}
}
\value{
invisible
}
\description{
Runs a single pipeline
}
