% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{schedule_validity_check}
\alias{schedule_validity_check}
\title{Checks the validity of a schedule}
\usage{
schedule_validity_check(schedule)
}
\arguments{
\item{schedule}{a schedule table returned from \code{build_schedule}}
}
\value{
invisible or error
}
\description{
Checks the validity of a schedule
}
